streamlit/requirements.txt
black
jupyterlab
jupyterlab_code_formatter
jupyterlab-git
lckr-jupyterlab-variableinspector
jupyterlab_widgets
ipywidgets
import-ipynb
numpy
pandas
scipy
requests
huggingface
langchain_chroma
sentence-transformers
langchain
langchain-community
langchain-openai 
python-dotenv
streamlit
altair<5